requirements.txt
pandas
numpy
requests
openpyxl
pymongo
mysql.connector
sqlalchemy